tests/requirements.txt
pytest==3.0.7
PyYAML==3.12
selenium==3.3.3
pprint
